requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
anthropic==0.7.8
chromadb==0.4.15
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
lxml==4.9.3
aiofiles==23.2.0
jinja2==3.1.2
python-multipart==0.0.6
